GET Request Conection Done, further tests required
Finished GET request connection in fingerprint-auth android app. Application receives the request and outputs it in the application as part of the test (The GET request needs further attention, because it doesn't output the values of the request but some other content). Refactored the code and closed the resource stream.