PUT Request Connection Done, further tests required
Creating data to update the object on the server. Initialization and configuration of the request with connection to the server. Updateing data on the server and checking if update successful or not. Closing resource streams. Further tests required because the object in the database doesn't change internal values but only updates ID & version of the object to newer version.